requirements.txt
typer==0.9.0
Pillow==10.0.1
